made separate app for events
