+++ eatpoint/api/serializers/establishments.py
@@ -22,7 +22,6 @@ from establishments.models import (
     ImageEstablishment,
     TypeEst,
     City,
-    Event,
 )
 from reservation.models import Availability
 from reviews.models import Review
@@ -433,74 +432,3 @@ class EstablishmentEditSerializer(serializers.ModelSerializer):
             instance,
             context={"request": self.context.get("request")},
         ).data
-
-
-class EventSerializer(serializers.ModelSerializer):
-    """Сериализатор событий"""
-
-    establishment = serializers.SlugRelatedField(
-        slug_field="name",
-        read_only=True,
-    )
-    image = Base64ImageField()
-
-    class Meta:
-        model = Event
-        fields = (
-            "name",
-            "establishment",
-            "description",
-            "date_start",
-            "date_end",
-            "type_event",
-            "price",
-            "image",
-        )
-
-
-class EventEditSerializer(serializers.ModelSerializer):
-    """Сериализатор событий"""
-
-    establishment = serializers.SlugRelatedField(
-        slug_field="name",
-        read_only=True,
-    )
-
-    class Meta:
-        model = Event
-        fields = (
-            "name",
-            "establishment",
-            "description",
-            "date_start",
-            "date_end",
-            "type_event",
-            "price",
-        )
-
-    def to_representation(self, instance):
-        return EventSerializer(
-            instance,
-            context={"request": self.context.get("request")},
-        ).data
-
-    def validate(self, data):
-        name = data.get("name")
-        date_start = data.get("date_start")
-        if Event.objects.filter(name=name, date_start=date_start).exists():
-            raise ValueError()
-
-    def create(self, validated_data):
-        type_event = validated_data.pop("type_event")
-        event = Event.objects.create(**validated_data)
-        event.kitchens.set(type_event)
-        return event
-
-    def update(self, instance, validated_data):
-        if "type_event" in validated_data:
-            instance.type_event.set(validated_data.pop("type_event"))
-
-        return super().update(
-            instance,
-            validated_data,
-        )
